+++ CHANGELOG.md
@@ -51,6 +51,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ([#892](https://github.com/nix-rust/nix/pull/892))
 - Remove `IFF_NOTRAILERS` on OpenBSD, as it has been removed in OpenBSD 6.3
   ([#893](https://github.com/nix-rust/nix/pull/893))
+- ControlMessage enum now supports SCM_TIMESTAMPING messages on Linux
+  ([#902](https://github.com/nix-rust/nix/pull/902))
 
 ### Fixed
 - Fixed possible panics when using `SigAction::flags` on Linux